Refactor: move things from experiments to more specific locations
