+++ src/evaluating_rewards/analysis/reward_figures/point_mass.py
